# swagger_server/models/paged_reviews.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.review import Review  # noqa: F401,E501
from swagger_server import util


class PagedReviews(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, items: List[Review]=None, page: int=None, page_size: int=None, total: int=None):  # noqa: E501
        """PagedReviews - a model defined in Swagger

        :param items: The items of this PagedReviews.  # noqa: E501
        :type items: List[Review]
        :param page: The page of this PagedReviews.  # noqa: E501
        :type page: int
        :param page_size: The page_size of this PagedReviews.  # noqa: E501
        :type page_size: int
        :param total: The total of this PagedReviews.  # noqa: E501
        :type total: int
        """
        self.swagger_types = {
            'items': List[Review],
            'page': int,
            'page_size': int,
            'total': int
        }

        self.attribute_map = {
            'items': 'items',
            'page': 'page',
            'page_size': 'pageSize',
            'total': 'total'
        }
        self._items = items
        self._page = page
        self._page_size = page_size
        self._total = total

    @classmethod
    def from_dict(cls, dikt) -> 'PagedReviews':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PagedReviews of this PagedReviews.  # noqa: E501
        :rtype: PagedReviews
        """
        return util.deserialize_model(dikt, cls)

    @property
    def items(self) -> List[Review]:
        """Gets the items of this PagedReviews.


        :return: The items of this PagedReviews.
        :rtype: List[Review]
        """
        return self._items

    @items.setter
    def items(self, items: List[Review]):
        """Sets the items of this PagedReviews.


        :param items: The items of this PagedReviews.
        :type items: List[Review]
        """

        self._items = items

    @property
    def page(self) -> int:
        """Gets the page of this PagedReviews.


        :return: The page of this PagedReviews.
        :rtype: int
        """
        return self._page

    @page.setter
    def page(self, page: int):
        """Sets the page of this PagedReviews.


        :param page: The page of this PagedReviews.
        :type page: int
        """

        self._page = page

    @property
    def page_size(self) -> int:
        """Gets the page_size of this PagedReviews.


        :return: The page_size of this PagedReviews.
        :rtype: int
        """
        return self._page_size

    @page_size.setter
    def page_size(self, page_size: int):
        """Sets the page_size of this PagedReviews.


        :param page_size: The page_size of this PagedReviews.
        :type page_size: int
        """

        self._page_size = page_size

    @property
    def total(self) -> int:
        """Gets the total of this PagedReviews.


        :return: The total of this PagedReviews.
        :rtype: int
        """
        return self._total

    @total.setter
    def total(self, total: int):
        """Sets the total of this PagedReviews.


        :param total: The total of this PagedReviews.
        :type total: int
        """

        self._total = total