# swagger_server/models/review.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.review_status import ReviewStatus  # noqa: F401,E501
from swagger_server import util


class Review(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, user_id: str=None, product_id: str=None, order_id: str=None, rating: int=None, title: str=None, body: str=None, status: ReviewStatus=None, created_at: datetime=None, updated_at: datetime=None):  # noqa: E501
        """Review - a model defined in Swagger

        :param id: The id of this Review.  # noqa: E501
        :type id: str
        :param user_id: The user_id of this Review.  # noqa: E501
        :type user_id: str
        :param product_id: The product_id of this Review.  # noqa: E501
        :type product_id: str
        :param order_id: The order_id of this Review.  # noqa: E501
        :type order_id: str
        :param rating: The rating of this Review.  # noqa: E501
        :type rating: int
        :param title: The title of this Review.  # noqa: E501
        :type title: str
        :param body: The body of this Review.  # noqa: E501
        :type body: str
        :param status: The status of this Review.  # noqa: E501
        :type status: ReviewStatus
        :param created_at: The created_at of this Review.  # noqa: E501
        :type created_at: datetime
        :param updated_at: The updated_at of this Review.  # noqa: E501
        :type updated_at: datetime
        """
        self.swagger_types = {
            'id': str,
            'user_id': str,
            'product_id': str,
            'order_id': str,
            'rating': int,
            'title': str,
            'body': str,
            'status': ReviewStatus,
            'created_at': datetime,
            'updated_at': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'user_id': 'userId',
            'product_id': 'productId',
            'order_id': 'orderId',
            'rating': 'rating',
            'title': 'title',
            'body': 'body',
            'status': 'status',
            'created_at': 'createdAt',
            'updated_at': 'updatedAt'
        }
        self._id = id
        self._user_id = user_id
        self._product_id = product_id
        self._order_id = order_id
        self._rating = rating
        self._title = title
        self._body = body
        self._status = status
        self._created_at = created_at
        self._updated_at = updated_at

    @classmethod
    def from_dict(cls, dikt) -> 'Review':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Review of this Review.  # noqa: E501
        :rtype: Review
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Review.


        :return: The id of this Review.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Review.


        :param id: The id of this Review.
        :type id: str
        """

        self._id = id

    @property
    def user_id(self) -> str:
        """Gets the user_id of this Review.


        :return: The user_id of this Review.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: str):
        """Sets the user_id of this Review.


        :param user_id: The user_id of this Review.
        :type user_id: str
        """

        self._user_id = user_id

    @property
    def product_id(self) -> str:
        """Gets the product_id of this Review.

        Merch u otros artículos  # noqa: E501

        :return: The product_id of this Review.
        :rtype: str
        """
        return self._product_id

    @product_id.setter
    def product_id(self, product_id: str):
        """Sets the product_id of this Review.

        Merch u otros artículos  # noqa: E501

        :param product_id: The product_id of this Review.
        :type product_id: str
        """

        self._product_id = product_id

    @property
    def order_id(self) -> str:
        """Gets the order_id of this Review.

        Pedido asociado que habilita la reseña  # noqa: E501

        :return: The order_id of this Review.
        :rtype: str
        """
        return self._order_id

    @order_id.setter
    def order_id(self, order_id: str):
        """Sets the order_id of this Review.

        Pedido asociado que habilita la reseña  # noqa: E501

        :param order_id: The order_id of this Review.
        :type order_id: str
        """

        self._order_id = order_id

    @property
    def rating(self) -> int:
        """Gets the rating of this Review.


        :return: The rating of this Review.
        :rtype: int
        """
        return self._rating

    @rating.setter
    def rating(self, rating: int):
        """Sets the rating of this Review.


        :param rating: The rating of this Review.
        :type rating: int
        """

        self._rating = rating

    @property
    def title(self) -> str:
        """Gets the title of this Review.


        :return: The title of this Review.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Review.


        :param title: The title of this Review.
        :type title: str
        """

        self._title = title

    @property
    def body(self) -> str:
        """Gets the body of this Review.


        :return: The body of this Review.
        :rtype: str
        """
        return self._body

    @body.setter
    def body(self, body: str):
        """Sets the body of this Review.


        :param body: The body of this Review.
        :type body: str
        """

        self._body = body

    @property
    def status(self) -> ReviewStatus:
        """Gets the status of this Review.


        :return: The status of this Review.
        :rtype: ReviewStatus
        """
        return self._status

    @status.setter
    def status(self, status: ReviewStatus):
        """Sets the status of this Review.


        :param status: The status of this Review.
        :type status: ReviewStatus
        """

        self._status = status

    @property
    def created_at(self) -> datetime:
        """Gets the created_at of this Review.


        :return: The created_at of this Review.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at: datetime):
        """Sets the created_at of this Review.


        :param created_at: The created_at of this Review.
        :type created_at: datetime
        """

        self._created_at = created_at

    @property
    def updated_at(self) -> datetime:
        """Gets the updated_at of this Review.


        :return: The updated_at of this Review.
        :rtype: datetime
        """
        return self._updated_at

    @updated_at.setter
    def updated_at(self, updated_at: datetime):
        """Sets the updated_at of this Review.


        :param updated_at: The updated_at of this Review.
        :type updated_at: datetime
        """

        self._updated_at = updated_at