# swagger_server/models/review_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ReviewStatus(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    """
    allowed enum values
    """
    VISIBLE = "visible"
    HIDDEN = "hidden"
    FLAGGED = "flagged"
    def __init__(self):  # noqa: E501
        """ReviewStatus - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'ReviewStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ReviewStatus of this ReviewStatus.  # noqa: E501
        :rtype: ReviewStatus
        """
        return util.deserialize_model(dikt, cls)