# swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Usuarios y Autorización
  description: ""
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/universidadpolitcnic-44f/Usuarios_y_autorizacion/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.test/v1
...
paths:
...
  /users/{userId}/followers:
    get:
      tags:
      - Users (admin)
      summary: Listar seguidores de un usuario (admin o dueño)
      operationId: users_user_id_followers_get
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: pageSize
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Seguidores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedUsers"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
...
  /admin/reviews:
    get:
      tags:
      - Users (admin)
      summary: Moderar reseñas (listar/filtrar)
      operationId: admin_reviews_get
      parameters:
      - name: userId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: productId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: status
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - visible
          - hidden
          - flagged
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: pageSize
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Reseñas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedReviews"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
  /admin/reviews/{reviewId}:
    patch:
      tags:
      - Users (admin)
      summary: Moderar estado de una reseña
      operationId: admin_reviews_review_id_patch
      parameters:
      - name: reviewId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reviews_reviewId_body"
        required: true
      responses:
        "200":
          description: Reseña moderada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
...
  /users:
    get:
      tags:
      - Users (admin)
      summary: Listado de usuarios
      operationId: users_get
      parameters:
      - name: q
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: role
        in: query
        required: false
        style: form
        explode: true
        schema:
          $ref: "#/components/schemas/Role"
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: pageSize
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Página de usuarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
  /users/{userId}:
    get:
      tags:
      - Users (admin)
      summary: Detalle de usuario
      operationId: users_user_id_get
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
    patch:
      tags:
      - Users (admin)
      summary: Edita usuario
      operationId: users_user_id_patch
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAdminUpdate"
        required: true
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
  /users/{userId}/roles:
    patch:
      tags:
      - Users (admin)
      summary: Asigna/quita roles
      operationId: users_user_id_roles_patch
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userId_roles_body"
        required: true
      responses:
        "204":
          description: Roles actualizados
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.users_(admin)controller
components:
  schemas:
    Role:
      type: string
      enum:
      - listener
      - artist
      - admin
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        avatarUrl:
          type: string
          format: uri
        bio:
          type: string
      example:
        role: listener
        avatarUrl: http://example.com/aeiou
        bio: bio
        id: id
        username: username
...
    UserAdminUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        disabled:
          type: boolean
...
    PagedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserPublic"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      example:
        total: 1
        pageSize: 6
        page: 0
        items:
        - role: listener
          avatarUrl: http://example.com/aeiou
          bio: bio
          id: id
          username: username
        - role: listener
          avatarUrl: http://example.com/aeiou
          bio: bio
          id: id
          username: username
...
    ReviewStatus:
      type: string
      enum:
      - visible
      - hidden
      - flagged
    Review:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productId:
          type: string
          description: Merch u otros artículos
        orderId:
          type: string
          description: Pedido asociado que habilita la reseña
        rating:
          maximum: 5
          minimum: 1
          type: integer
        title:
          maxLength: 120
          type: string
        body:
          maxLength: 2000
          type: string
        status:
          $ref: "#/components/schemas/ReviewStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        productId: productId
        orderId: orderId
        rating: 1
        id: id
        title: title
        body: body
        userId: userId
        status: visible
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    PagedReviews:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      example:
        total: 5
        pageSize: 1
        page: 6
        items:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          productId: productId
          orderId: orderId
          rating: 1
          id: id
          title: title
          body: body
          userId: userId
          status: visible
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          productId: productId
          orderId: orderId
          rating: 1
          id: id
          title: title
          body: body
          userId: userId
          status: visible
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    reviews_reviewId_body:
      type: object
      properties:
        status:
          type: string
          enum:
          - visible
          - hidden
          - flagged
...
    inline_response_200:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserPublic"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      example:
        total: 1
        pageSize: 6
        page: 0
        items:
        - role: listener
          avatarUrl: http://example.com/aeiou
          bio: bio
          id: id
          username: username
        - role: listener
          avatarUrl: http://example.com/aeiou
          bio: bio
          id: id
          username: username
    userId_roles_body:
      required:
      - roles
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth